requirements.txt
python-igraph==0.9.6
annoy==1.17.0
pandas==1.2.4
numpy==1.20.3
tqdm==4.61.1
torch==1.8.0
transformers==4.6.1
sentence-transformers==1.2.0
scikit-learn==0.24.2